packet: proper return type for vlan_tci_to_pcp()
The return type of vlan_tci_to_pcp() was int where it's expected to be
uint_8 and causing implicit truncation when the function is used. On
some platforms such as macOS, where PRIu8 is defined as "hhx" and no
promotion of short to int is done, the compiler might throw out Wformat
message for ds_put_format() calls on the returns value of
vlan_tci_to_pcp().

vlan_tci_to_cfi() is also fixed with unit_8 as return type although the
function is not currently being used anywhere.

Format strings in ds_put_format() for printing out returned values from
vlan_tci_to_pcp() were updated to ensure PRIu8 or PRIx8 are used for
portability.

Signed-off-by: Shu Shen <shu.shen@gmail.com>