packet: proper return type for vlan_tci_to_pcp()
The return type of vlan_tci_to_pcp() was int where it's expected to be
uint8_t and causing implicit truncation when the function is used.

On some platforms such as macOS, where PRIu8 is defined as "hhx" and no
promotion of short to int is done, the compiler might throw out Wformat
message for ds_put_format() calls on the returns value of
vlan_tci_to_pcp().

vlan_tci_to_cfi() is also fixed with uint8_t as return type although the
function is not currently being used anywhere.

Signed-off-by: Shu Shen <shu.shen@gmail.com>